refactor(shared): file-organization sweep over app/shared/models/db
The three admin action input variants each take their own model file with their own schema, the base shape they extend becomes one too, and ExecuteAdminActionInput keeps the union and composes the discriminated union from the variant schemas — the shape the sheet column family already has. Every other multi-export file in the tree is a type beside its schema, an enum beside its values array, or an entity beside its property names.
