+++ apps/web/shared/models/db/moderation/BaseExecuteAdminActionInput.ts
@@ -1,0 +1,14 @@
+import type { RoomInMessage, User } from "@esposter/db-schema";
+
+import { roomIdSchema, selectUserSchema } from "@esposter/db-schema";
+import { z } from "zod";
+
+export interface BaseExecuteAdminActionInput {
+  roomId: RoomInMessage["id"];
+  targetUserId: User["id"];
+}
+
+export const baseExecuteAdminActionInputSchema = z.object({
+  ...roomIdSchema.shape,
+  targetUserId: selectUserSchema.shape.id,
+}) satisfies z.ZodType<BaseExecuteAdminActionInput>;
